+++ Changelog.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+09/11/2022
+### Added
+- Enrollment Page
+- New components (Select, Input, Date)
+
 09/08/2022
 ### Added
 - Mechanics Page